add model test1
